requirements.txt
africastalking>=1.2.7
aiohttp>=3.9.5
aiosignal>=1.3.1
annotated-types>=0.7.0
arrow>=1.3.0
asgiref>=3.8.1
attrs>=23.2.0
binaryornot>=0.4.4
cent>=5.0.0
certifi>=2024.6.2
cffi>=1.16.0
chardet>=5.2.0
charset-normalizer>=3.3.2
click>=8.1.7
cookiecutter>=2.6.0
crispy-bootstrap5>=2024.2
cryptography>=42.0.7
Django>=5.0.6
django-allauth>=0.63.3
django-allauth-ui>=1.2.0
django-browser-reload>=1.12.1
django-crispy-forms>=2.1
django-instant>=4.2.0
django-phonenumber-field>=7.3.0
django-phonenumbers>=1.0.1
django-widget-tweaks>=1.5.0
et-xmlfile>=1.1.0
frozenlist>=1.4.1
idna>=3.7
Jinja2>=3.1.4
jwt>=1.3.1
markdown-it-py>=3.0.0
MarkupSafe>=2.1.5
mdurl>=0.1.2
multidict>=6.0.5
numpy>=1.26.4
openpyxl>=3.1.3
pandas>=2.2.2
phonenumbers>=8.13.40
pycparser>=2.22
pydantic>=2.7.2
pydantic_core>=2.18.3
Pygments>=2.18.0
PyJWT>=2.8.0
python-dateutil>=2.9.0.post0
python-decouple>=3.8
python-slugify>=8.0.4
pytz>=2024.1
PyYAML>=6.0.1
requests>=2.32.3
rich>=13.7.1
schema>=0.7.7
six>=1.16.0
slippers>=0.6.2
sqlparse>=0.5.0
text-unidecode>=1.3
typeguard>=2.13.3
types-python-dateutil>=2.9.0.20240316
types-requests>=2.32.0.20240602
typing_extensions>=4.12.1
tzdata>=2024.1
urllib3>=2.2.1
yarl>=1.9.4
